Move entities.json into markup5ever
This commit contains:

- Extracting the file itself
- Extracting build.rs details from html5ever and xml5ever into
markup5ever
- Updating markup5ever dependencies
- Updating xml/html 5ever files to use markup5ever
