+++ markup5ever/Cargo.toml
@@ -16,8 +16,13 @@ heap_size = ["heapsize", "heapsize_derive", "string_cache/heapsize"]
 
 [dependencies]
 string_cache = "0.5"
+phf = "0.7"
 heapsize = { version = "0.3", optional = true }
 heapsize_derive = { version = "0.1", optional = true }
 
 [build-dependencies]
 string_cache_codegen = "0.4"
+phf_codegen = "0.7.3"
+rustc-serialize = "0.3.15"
+syn = { version = "0.11", features = ["full", "visit"] }
+
