+++ xml5ever/src/tokenizer/char_ref/mod.rs
@@ -19,7 +19,8 @@ use std::borrow::Cow::Borrowed;
 pub use self::Status::*;
 use self::State::*;
 
-mod data;
+use markup5ever::data;
+
 
 //§ tokenizing-character-references
 pub struct CharRef {
